pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yodawg"
version = "0.1.0"
description = "Download wallpapers from Reddit and generate terminal color schemes"
dependencies = ["requests", "aiohttp", "pillow", "pywal16", "colorz"]

[project.scripts]
reddit = "yodawg.reddit:main"
wal = "pywal.__main__:main"
rice = "yodawg.rice:main"
save-wallpaper = "yodawg.savewall:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
